Update README.md
Nowadays, there is an urgent necessity to detect and classify drone swarms, both for military and civilian purposes. One of the most promising methods is detection of drones using a radar signal. It has been proved that RF signatures along with convolutional neural networks (CNNs) provide an accurate approach for classification of the single drones. This repository presents the Python scripts to train, validate and test a deep learning model for classification of drone swarms.
CNN is applied to Short-Time Fourier Transform (STFT) spectrograms of radar signals that are reflected from a drone swarm due to propellers movement. To train and demonstrate the capabilities of the applied approach, a synthetic dataset has been generated, in which the radar return is described by the mathematical model of Martin and Mulgrew. Realistic parameters of drones and radar are taken from paper
Divy Raval, Emily Hunter, Sinclair Hudson, Anthony Damini, and Bhashyam Balaji, Convolutional Neural Networks for Classification of Drones Using Radars, Drones 2021, 5, 149. https://doi.org/10.3390/drones5040149.